refactor(arch): relocate dynamic_forms from features/ to shared/ (#121)
dynamic_forms is not a feature — it is a forms engine. It has no
domain entity of its own (FormSchemaEntity/FormFieldEntity describe
HOW to render data, not a business concept), and every actual
consumer is another feature (users, dashboard, future entities).
Its placement under features/ forced cross-feature imports which
the architecture guard test had to absorb as known exceptions —
6 entries just for this PR, with each new consumer adding more.

Moving the module to shared/dynamic_forms/ matches its true role:
a feature-agnostic, reusable subsystem consumed by features. The
features → shared import direction is allowed by default, so all
6 known-exception entries are removed in this commit.

What changed:
- lib/features/dynamic_forms/  → lib/shared/dynamic_forms/  (16 files)
- test/features/dynamic_forms/ → test/shared/dynamic_forms/ (11 files)
- All package: imports updated from features/dynamic_forms/* to
  shared/dynamic_forms/* across:
  * app/di/{app_dependencies,app_scope}.dart
  * app/router/app_router.dart
  * features/users/navigation/users_routes.dart
  * features/users/presentation/pages/user_extended_info_page.dart
  * test/mocks/mock_classes.dart
  * intra-module imports inside dynamic_forms
- test/architecture/import_guard_test.dart: removed the 6 now-dead
  cross-feature exceptions for users → dynamic_forms.

Verification:
- fvm dart analyze: No issues found
- fvm dart format --line-length=120 --set-exit-if-changed .: 0 changed
- fvm flutter test: 1435 tests pass
- Architecture guard test: passes naturally (features → shared is
  the allowed direction)
