--- test/features/dynamic_forms/application/usecases/load_form_bundle_usecase_test.dart
@@ -1,53 +1,0 @@
-import 'package:flutter_bloc_advance/core/errors/app_error.dart';
-import 'package:flutter_bloc_advance/core/result/result.dart';
-import 'package:flutter_bloc_advance/features/dynamic_forms/application/usecases/load_form_bundle_usecase.dart';
-import 'package:flutter_bloc_advance/features/dynamic_forms/domain/entities/form_bundle_entity.dart';
-import 'package:flutter_bloc_advance/features/dynamic_forms/domain/entities/form_schema_entity.dart';
-import 'package:flutter_bloc_advance/features/dynamic_forms/domain/repositories/dynamic_form_repository.dart';
-import 'package:flutter_test/flutter_test.dart';
-
-class _FakeRepo implements IDynamicFormRepository {
-  Result<FormBundleEntity>? bundleResult;
-  String? lastEndpoint;
-
-  @override
-  Future<Result<FormSchemaEntity>> fetchSchema(String formId) async => const Failure(UnknownError('not used'));
-
-  @override
-  Future<Result<FormBundleEntity>> fetchBundle(String endpoint) async {
-    lastEndpoint = endpoint;
-    return bundleResult ?? const Failure(UnknownError('not configured'));
-  }
-
-  @override
-  Future<Result<String?>> submit(FormSubmitAction action, Map<String, dynamic> data) async =>
-      const Failure(UnknownError('not used'));
-}
-
-void main() {
-  group('LoadFormBundleUseCase', () {
-    test('forwards the endpoint to the repository and returns its Result', () async {
-      final bundle = const FormBundleEntity(
-        schema: FormSchemaEntity(id: 'extended', title: 'Extended Info'),
-        values: {'name': 'Alice', 'email': 'alice@example.com'},
-      );
-      final repo = _FakeRepo()..bundleResult = Success(bundle);
-      final useCase = LoadFormBundleUseCase(repo);
-
-      final result = await useCase('/admin/users/1/extended');
-
-      expect(result, isA<Success<FormBundleEntity>>());
-      expect(repo.lastEndpoint, '/admin/users/1/extended');
-    });
-
-    test('propagates a repository Failure unchanged', () async {
-      final repo = _FakeRepo()..bundleResult = const Failure(NetworkError('offline'));
-      final useCase = LoadFormBundleUseCase(repo);
-
-      final result = await useCase('/x');
-
-      expect(result, isA<Failure<FormBundleEntity>>());
-      expect((result as Failure).error, isA<NetworkError>());
-    });
-  });
-}
